backend/rag/requirements.txt
python-dotenv
PyMuPDF
python-docx
qdrant-client
google-generativeai
langchain-google-genai
langchain-core
langchain-text-splitters
